PlayList App C++ 2014
Written in C++ in 2014. This program loads a list of songs from an
external file and stores them in a hash table adt. Each song includes a
title, an artist, an album name, the number
of minutes, and how many
times it has been played. The user can enter a key word and retrieve all
songs in the table with that word in the title. The user can also delete
a song by key word and then chose one of the songs from the results to
delete.*/